refactor: move remaining providers to tensorzero-providers (step 2)
Move all remaining provider files that depend on openai/:
- openai/, fireworks/
- azure, deepseek, groq, hyperbolic, mistral, openrouter, sglang,
  tgi, together, vllm, xai

Also:
- Relocate orphan-rule From impls (Stored* ↔ provider types) to providers
- Update optimizer row types to own FunctionToolDef (serialize_with pattern)
- Add tool_serializer module in tensorzero-optimizers
- Remove From impls from core's model.rs and openai_rft/mod.rs

tensorzero-core/src/providers/mod.rs is now a pure re-export shim
(plus dummy.rs and test_helpers.rs which stay in core).
