+++ crates/tensorzero-core/src/providers/mod.rs
@@ -1,26 +1,13 @@
 // Re-exported from tensorzero-providers:
 pub use tensorzero_providers::{
-    anthropic, aws_bedrock, aws_common, aws_sagemaker, chat_completions, gcp_vertex_anthropic,
-    google_ai_studio_gemini, helpers, helpers_thinking_block,
+    anthropic, aws_bedrock, aws_common, aws_sagemaker, azure, chat_completions, deepseek,
+    fireworks, gcp_vertex_anthropic, google_ai_studio_gemini, groq, helpers,
+    helpers_thinking_block, hyperbolic, mistral, openai, openrouter, sglang, tgi, together, vllm,
+    xai,
 };
 
 // gcp_vertex_gemini is overridden locally to add the core-only `optimization` submodule.
 pub mod gcp_vertex_gemini;
-
-// Still local (depend on openai/ which hasn't moved yet):
-pub mod azure;
-pub mod deepseek;
-pub mod fireworks;
-pub mod groq;
-pub mod hyperbolic;
-pub mod mistral;
-pub mod openai;
-pub mod openrouter;
-pub mod sglang;
-pub mod tgi;
-pub mod together;
-pub mod vllm;
-pub mod xai;
 
 #[cfg(any(test, feature = "e2e_tests"))]
 pub mod dummy;
