+++ crates/tensorzero-providers/src/groq.rs
@@ -13,34 +13,34 @@ use tensorzero_types_providers::groq::{
 };
 use tokio::time::Instant;
 
-use crate::endpoints::inference::InferenceCredentials;
-use crate::error::{
+use tensorzero_error::{
     DelayedError, DisplayOrDebugGateway, Error, ErrorDetails, warn_discarded_thought_block,
 };
-use crate::http::TensorzeroHttpClient;
-use crate::inference::types::batch::{BatchRequestRow, PollBatchInferenceResponse};
-use crate::inference::types::chat_completion_inference_params::{
-    ChatCompletionInferenceParamsV2, ServiceTier, warn_inference_parameter_not_supported,
-};
-use crate::inference::types::usage::raw_usage_entries_from_value;
-use crate::inference::types::{ApiType, FinishReason, ProviderInferenceResponseStreamInner};
-use crate::inference::types::{
-    ContentBlock, ContentBlockChunk, ContentBlockOutput, Latency, ModelInferenceRequest,
-    ModelInferenceRequestJsonMode, ObjectStorageFile, PeekableProviderInferenceResponseStream,
-    ProviderInferenceResponse, ProviderInferenceResponseArgs, ProviderInferenceResponseChunk,
-    RequestMessage, Role, Text, TextChunk, Thought, ThoughtChunk, Unknown, Usage,
-    batch::StartBatchProviderInferenceResponse,
-    resolved_input::{FileUrl, LazyFile, LazyFileExt},
-};
-use crate::inference::{InferenceProvider, TensorZeroEventError};
-use crate::model::Credential;
-use crate::model::{ModelProviderRequestInfo, ProviderInferenceRequest};
-use crate::tool::{ToolCall, ToolCallChunk, ToolChoice};
+use tensorzero_http::TensorzeroHttpClient;
 use tensorzero_inference_types::FunctionToolDef;
+use tensorzero_inference_types::ToolCallChunk;
+use tensorzero_inference_types::credentials::Credential;
+use tensorzero_inference_types::credentials::{ModelProviderRequestInfo, ProviderInferenceRequest};
+use tensorzero_inference_types::provider_trait::{InferenceProvider, TensorZeroEventError};
+use tensorzero_inference_types::raw_usage_entries_from_value;
+use tensorzero_inference_types::utils::warn_inference_parameter_not_supported;
+use tensorzero_inference_types::{BatchRequestRow, PollBatchInferenceResponse};
+use tensorzero_inference_types::{
+    ContentBlock, ContentBlockChunk, ContentBlockOutput, FileUrl, Latency, LazyFile, LazyFileExt,
+    ModelInferenceRequest, ModelInferenceRequestJsonMode, PeekableProviderInferenceResponseStream,
+    ProviderInferenceResponse, ProviderInferenceResponseArgs, ProviderInferenceResponseChunk,
+    RequestMessage, StartBatchProviderInferenceResponse, TextChunk, ThoughtChunk, Usage,
+};
+use tensorzero_inference_types::{FinishReason, ProviderInferenceResponseStreamInner};
+use tensorzero_types::ApiType;
+use tensorzero_types::inference_params::InferenceCredentials;
+use tensorzero_types::inference_params::{ChatCompletionInferenceParamsV2, ServiceTier};
+use tensorzero_types::{ObjectStorageFile, Role, Text, Thought, Unknown};
+use tensorzero_types::{ToolCall, ToolChoice};
 use uuid::Uuid;
 
-use crate::providers::chat_completions::prepare_chat_completion_tools;
-use crate::providers::helpers::{
+use crate::chat_completions::prepare_chat_completion_tools;
+use crate::helpers::{
     convert_stream_error, inject_extra_request_data_and_send,
     inject_extra_request_data_and_send_eventsource, warn_cannot_forward_url_if_missing_mime_type,
 };
@@ -1476,19 +1476,17 @@ mod tests {
     use serde_json::json;
 
     use super::*;
-    use crate::inference::types::file::Detail;
-    use crate::inference::types::resolved_input::LazyFile;
-    use crate::inference::types::storage::{StorageKind, StoragePath};
-    use crate::inference::types::{
-        ContentBlock, FunctionType, ObjectStorageFile, ObjectStoragePointer,
-        PendingObjectStoreFile, RequestMessage,
+    use crate::test_helpers::capture_logs;
+    use crate::test_helpers::{
+        MULTI_PROVIDER_TOOL_CONFIG, QUERY_TOOL_DEF, WEATHER_PROVIDER_TOOL_CONFIG, WEATHER_TOOL_DEF,
     };
-    use crate::providers::test_helpers::{
-        MULTI_PROVIDER_TOOL_CONFIG, QUERY_TOOL, WEATHER_PROVIDER_TOOL_CONFIG, WEATHER_TOOL,
-    };
-    use crate::tool::ToolCallConfig;
-    use crate::utils::testing::capture_logs;
+    use tensorzero_inference_types::LazyFile;
     use tensorzero_inference_types::ProviderToolCallConfig;
+    use tensorzero_inference_types::{ContentBlock, PendingObjectStoreFile, RequestMessage};
+    use tensorzero_types::Detail;
+    use tensorzero_types::ObjectStoragePointer;
+    use tensorzero_types::{FunctionType, ObjectStorageFile};
+    use tensorzero_types::{StorageKind, StoragePath};
     use tensorzero_types_providers::groq::{
         GroqChatChunkChoice, GroqDelta, GroqFunctionCallChunk, GroqResponseFunctionCall,
         GroqResponseMessage, GroqToolCallChunk,
@@ -1702,14 +1700,14 @@ mod tests {
         );
         assert!(groq_request.tools.is_some());
         let tools = groq_request.tools.as_ref().unwrap();
-        assert_eq!(tools[0].function.name, WEATHER_TOOL.name());
-        assert_eq!(tools[0].function.parameters, WEATHER_TOOL.parameters());
+        assert_eq!(tools[0].function.name, WEATHER_TOOL_DEF.name);
+        assert_eq!(tools[0].function.parameters, &WEATHER_TOOL_DEF.parameters);
         assert_eq!(
             groq_request.tool_choice,
             Some(GroqToolChoice::Specific(SpecificToolChoice {
                 r#type: GroqToolType::Function,
                 function: SpecificToolFunction {
-                    name: WEATHER_TOOL.name(),
+                    name: &WEATHER_TOOL_DEF.name,
                 }
             }))
         );
@@ -2165,10 +2163,10 @@ mod tests {
             prepare_groq_tools(&request_with_tools).unwrap();
         let tools = tools.unwrap();
         assert_eq!(tools.len(), 2);
-        assert_eq!(tools[0].function.name, WEATHER_TOOL.name());
-        assert_eq!(tools[0].function.parameters, WEATHER_TOOL.parameters());
-        assert_eq!(tools[1].function.name, QUERY_TOOL.name());
-        assert_eq!(tools[1].function.parameters, QUERY_TOOL.parameters());
+        assert_eq!(tools[0].function.name, WEATHER_TOOL_DEF.name);
+        assert_eq!(tools[0].function.parameters, &WEATHER_TOOL_DEF.parameters);
+        assert_eq!(tools[1].function.name, QUERY_TOOL_DEF.name);
+        assert_eq!(tools[1].function.parameters, &QUERY_TOOL_DEF.parameters);
         let tool_choice = tool_choice.unwrap();
         assert_eq!(
             tool_choice,
@@ -2176,12 +2174,12 @@ mod tests {
         );
         let parallel_tool_calls = parallel_tool_calls.unwrap();
         assert!(parallel_tool_calls);
-        let tool_config = ToolCallConfig {
+        let tool_config = ProviderToolCallConfig {
             tool_choice: ToolChoice::Required,
             parallel_tool_calls: Some(true),
             ..Default::default()
         };
-        let provider_tool_config = ProviderToolCallConfig::from(&tool_config);
+        let provider_tool_config = tool_config.clone();
 
         // Test no tools but a tool choice and make sure tool choice output is None
         let request_without_tools = ModelInferenceRequest {
@@ -2214,23 +2212,24 @@ mod tests {
 
     #[test]
     fn test_prepare_groq_tools_with_allowed_tools() {
-        use crate::tool::{AllowedTools, AllowedToolsChoice};
+        use tensorzero_inference_types::{AllowedTools, AllowedToolsChoice};
 
         // Test with allowed_tools specified
-        let tool_config = ToolCallConfig {
-            static_tools_available: vec![WEATHER_TOOL.clone(), QUERY_TOOL.clone()],
-            dynamic_tools_available: vec![],
+        let tool_config = ProviderToolCallConfig {
+            tools: vec![WEATHER_TOOL_DEF.clone(), QUERY_TOOL_DEF.clone()],
             provider_tools: vec![],
             openai_custom_tools: vec![],
             tool_choice: ToolChoice::Auto,
             parallel_tool_calls: Some(false),
             allowed_tools: AllowedTools {
-                tools: vec![WEATHER_TOOL.name().to_string()].into_iter().collect(),
+                tools: vec![WEATHER_TOOL_DEF.name.to_string()]
+                    .into_iter()
+                    .collect(),
                 choice: AllowedToolsChoice::Explicit,
             },
         };
 
-        let provider_tool_config = ProviderToolCallConfig::from(&tool_config);
+        let provider_tool_config = tool_config.clone();
         let request = ModelInferenceRequest {
             inference_id: uuid::Uuid::now_v7(),
             messages: vec![RequestMessage {
@@ -2271,7 +2270,7 @@ mod tests {
                 assert_eq!(allowed_tools_choice.allowed_tools.tools.len(), 1);
                 match &allowed_tools_choice.allowed_tools.tools[0] {
                     ToolReference::Function { function } => {
-                        assert_eq!(function.name, WEATHER_TOOL.name());
+                        assert_eq!(function.name, WEATHER_TOOL_DEF.name);
                     }
                     ToolReference::Custom { .. } => panic!("Expected Function variant"),
                 }
@@ -2874,7 +2873,7 @@ mod tests {
 
     #[test]
     fn test_groq_apply_inference_params_called() {
-        let logs_contain = crate::utils::testing::capture_logs();
+        let logs_contain = crate::test_helpers::capture_logs();
         let inference_params = ChatCompletionInferenceParamsV2 {
             reasoning_effort: Some("high".to_string()),
             service_tier: None,
