+++ crates/tensorzero-providers/src/lib.rs
@@ -2,12 +2,25 @@ pub mod anthropic;
 pub mod aws_bedrock;
 pub mod aws_common;
 pub mod aws_sagemaker;
+pub mod azure;
 pub mod chat_completions;
+pub mod deepseek;
+pub mod fireworks;
 pub mod gcp_vertex_anthropic;
 pub mod gcp_vertex_gemini;
 pub mod google_ai_studio_gemini;
+pub mod groq;
 pub mod helpers;
 pub mod helpers_thinking_block;
+pub mod hyperbolic;
+pub mod mistral;
+pub mod openai;
+pub mod openrouter;
+pub mod sglang;
+pub mod tgi;
+pub mod together;
+pub mod vllm;
+pub mod xai;
 
 #[cfg(test)]
 pub mod test_helpers;
