+++ crates/tensorzero-providers/src/mistral.rs
@@ -1,9 +1,6 @@
 use std::{borrow::Cow, time::Duration};
 
-use crate::{
-    http::{TensorZeroEventSource, TensorzeroHttpClient},
-    providers::openai::{OpenAIMessagesConfig, ReasoningFieldName},
-};
+use crate::openai::{OpenAIMessagesConfig, ReasoningFieldName};
 use futures::StreamExt;
 use lazy_static::lazy_static;
 use reqwest::StatusCode;
@@ -11,6 +8,7 @@ use reqwest_sse_stream::Event;
 use secrecy::{ExposeSecret, SecretString};
 use serde::Serialize;
 use serde_json::Value;
+use tensorzero_http::{TensorZeroEventSource, TensorzeroHttpClient};
 use tensorzero_types_providers::mistral::{
     MistralChatChunk, MistralContent, MistralContentChunk, MistralFinishReason, MistralResponse,
     MistralResponseChoice, MistralResponseFormat, MistralResponseToolCall, MistralThinkingSubChunk,
@@ -19,35 +17,27 @@ use tensorzero_types_providers::mistral::{
 use tokio::time::Instant;
 use url::Url;
 
-use crate::inference::types::usage::raw_usage_entries_from_value;
-use crate::{
-    endpoints::inference::InferenceCredentials,
-    error::{DelayedError, DisplayOrDebugGateway, Error, ErrorDetails},
-    inference::{
-        InferenceProvider,
-        types::{
-            ApiType, ContentBlock, ContentBlockChunk, ContentBlockOutput, FinishReason, Latency,
-            ModelInferenceRequest, ModelInferenceRequestJsonMode,
-            PeekableProviderInferenceResponseStream, ProviderInferenceResponse,
-            ProviderInferenceResponseArgs, ProviderInferenceResponseChunk,
-            ProviderInferenceResponseStreamInner, RequestMessage, Role, TextChunk, Thought,
-            ThoughtChunk, Usage,
-            batch::{
-                BatchRequestRow, PollBatchInferenceResponse, StartBatchProviderInferenceResponse,
-            },
-            chat_completion_inference_params::{
-                ChatCompletionInferenceParamsV2, warn_inference_parameter_not_supported,
-            },
-        },
-    },
-    model::{Credential, ModelProviderRequestInfo, ProviderInferenceRequest},
-    providers::helpers::{
-        check_new_tool_call_name, convert_stream_error, inject_extra_request_data_and_send,
-        inject_extra_request_data_and_send_eventsource,
-    },
-    tool::{ToolCall, ToolCallChunk, ToolChoice},
+use crate::helpers::{
+    check_new_tool_call_name, convert_stream_error, inject_extra_request_data_and_send,
+    inject_extra_request_data_and_send_eventsource,
 };
+use tensorzero_error::{DelayedError, DisplayOrDebugGateway, Error, ErrorDetails};
 use tensorzero_inference_types::FunctionToolDef;
+use tensorzero_inference_types::credentials::{
+    Credential, ModelProviderRequestInfo, ProviderInferenceRequest,
+};
+use tensorzero_inference_types::provider_trait::InferenceProvider;
+use tensorzero_inference_types::utils::warn_inference_parameter_not_supported;
+use tensorzero_inference_types::{
+    BatchRequestRow, ContentBlock, ContentBlockChunk, ContentBlockOutput, FinishReason, Latency,
+    ModelInferenceRequest, ModelInferenceRequestJsonMode, PeekableProviderInferenceResponseStream,
+    PollBatchInferenceResponse, ProviderInferenceResponse, ProviderInferenceResponseArgs,
+    ProviderInferenceResponseChunk, ProviderInferenceResponseStreamInner, RequestMessage,
+    StartBatchProviderInferenceResponse, TextChunk, ThoughtChunk, ToolCallChunk, Usage,
+    raw_usage_entries_from_value,
+};
+use tensorzero_types::inference_params::{ChatCompletionInferenceParamsV2, InferenceCredentials};
+use tensorzero_types::{ApiType, Role, Thought, ToolCall, ToolChoice};
 use uuid::Uuid;
 
 use super::openai::{
@@ -1078,10 +1068,11 @@ mod tests {
     use super::*;
     use googletest::prelude::*;
 
-    use crate::inference::types::{FunctionType, RequestMessage, Role};
-    use crate::providers::test_helpers::{QUERY_TOOL, WEATHER_PROVIDER_TOOL_CONFIG, WEATHER_TOOL};
-    use crate::tool::{AllowedTools, ToolCallConfig};
+    use crate::test_helpers::{QUERY_TOOL_DEF, WEATHER_PROVIDER_TOOL_CONFIG, WEATHER_TOOL_DEF};
+    use tensorzero_inference_types::AllowedTools;
     use tensorzero_inference_types::ProviderToolCallConfig;
+    use tensorzero_inference_types::RequestMessage;
+    use tensorzero_types::{FunctionType, Role};
     use tensorzero_types_providers::mistral::{
         MistralChatChunkChoice, MistralDelta, MistralResponseFunctionCall, MistralResponseMessage,
         MistralThinkingSubChunk,
@@ -1127,8 +1118,8 @@ mod tests {
         assert!(mistral_request.tools.is_some());
         let tools = mistral_request.tools.as_ref().unwrap();
         assert_eq!(tools.len(), 1);
-        assert_eq!(tools[0].function.name, WEATHER_TOOL.name());
-        assert_eq!(tools[0].function.parameters, WEATHER_TOOL.parameters());
+        assert_eq!(tools[0].function.name, WEATHER_TOOL_DEF.name);
+        assert_eq!(tools[0].function.parameters, &WEATHER_TOOL_DEF.parameters);
         assert_eq!(
             mistral_request.tool_choice,
             Some(MistralToolChoice::Specific(MistralSpecificToolChoice {
@@ -1142,23 +1133,24 @@ mod tests {
 
     #[test]
     fn test_prepare_mistral_tools_with_allowed_tools() {
-        use crate::tool::{AllowedTools, AllowedToolsChoice};
+        use tensorzero_inference_types::{AllowedTools, AllowedToolsChoice};
 
         // Test with allowed_tools specified - Mistral doesn't support allowed_tools constraint
-        let tool_config = ToolCallConfig {
-            static_tools_available: vec![WEATHER_TOOL.clone(), QUERY_TOOL.clone()],
-            dynamic_tools_available: vec![],
+        let tool_config = ProviderToolCallConfig {
+            tools: vec![WEATHER_TOOL_DEF.clone(), QUERY_TOOL_DEF.clone()],
             provider_tools: vec![],
             openai_custom_tools: vec![],
             tool_choice: ToolChoice::Auto,
             parallel_tool_calls: Some(false),
             allowed_tools: AllowedTools {
-                tools: vec![WEATHER_TOOL.name().to_string()].into_iter().collect(),
+                tools: vec![WEATHER_TOOL_DEF.name.to_string()]
+                    .into_iter()
+                    .collect(),
                 choice: AllowedToolsChoice::Explicit,
             },
         };
 
-        let provider_tool_config = ProviderToolCallConfig::from(&tool_config);
+        let provider_tool_config = tool_config.clone();
         let request = ModelInferenceRequest {
             inference_id: Uuid::now_v7(),
             messages: vec![RequestMessage {
@@ -1186,7 +1178,7 @@ mod tests {
         // Verify only allowed tools are returned (strict_tools_available respects allowed_tools)
         let tools = tools.unwrap();
         assert_eq!(tools.len(), 1);
-        assert_eq!(tools[0].function.name, WEATHER_TOOL.name());
+        assert_eq!(tools[0].function.name, WEATHER_TOOL_DEF.name);
 
         // Verify tool_choice
         let tool_choice = tool_choice.unwrap();
@@ -1202,9 +1194,8 @@ mod tests {
     #[test]
     fn test_prepare_mistral_tools_auto_mode() {
         // Test Auto mode with default allowed_tools
-        let tool_config = ToolCallConfig {
-            static_tools_available: vec![WEATHER_TOOL.clone(), QUERY_TOOL.clone()],
-            dynamic_tools_available: vec![],
+        let tool_config = ProviderToolCallConfig {
+            tools: vec![WEATHER_TOOL_DEF.clone(), QUERY_TOOL_DEF.clone()],
             provider_tools: vec![],
             openai_custom_tools: vec![],
             tool_choice: ToolChoice::Auto,
@@ -1212,7 +1203,7 @@ mod tests {
             allowed_tools: AllowedTools::default(),
         };
 
-        let provider_tool_config = ProviderToolCallConfig::from(&tool_config);
+        let provider_tool_config = tool_config.clone();
         let request = ModelInferenceRequest {
             inference_id: Uuid::now_v7(),
             messages: vec![RequestMessage {
@@ -1254,11 +1245,10 @@ mod tests {
 
     #[test]
     fn test_prepare_mistral_tools_required_mode() {
-        use crate::tool::AllowedTools;
+        use tensorzero_inference_types::AllowedTools;
 
-        let tool_config = ToolCallConfig {
-            static_tools_available: vec![WEATHER_TOOL.clone()],
-            dynamic_tools_available: vec![],
+        let tool_config = ProviderToolCallConfig {
+            tools: vec![WEATHER_TOOL_DEF.clone()],
             provider_tools: vec![],
             openai_custom_tools: vec![],
             tool_choice: ToolChoice::Required,
@@ -1266,7 +1256,7 @@ mod tests {
             allowed_tools: AllowedTools::default(),
         };
 
-        let provider_tool_config = ProviderToolCallConfig::from(&tool_config);
+        let provider_tool_config = tool_config.clone();
         let request = ModelInferenceRequest {
             inference_id: Uuid::now_v7(),
             messages: vec![RequestMessage {
@@ -1308,9 +1298,8 @@ mod tests {
 
     #[test]
     fn test_prepare_mistral_tools_none_mode() {
-        let tool_config = ToolCallConfig {
-            static_tools_available: vec![WEATHER_TOOL.clone()],
-            dynamic_tools_available: vec![],
+        let tool_config = ProviderToolCallConfig {
+            tools: vec![WEATHER_TOOL_DEF.clone()],
             provider_tools: vec![],
             openai_custom_tools: vec![],
             tool_choice: ToolChoice::None,
@@ -1318,7 +1307,7 @@ mod tests {
             allowed_tools: AllowedTools::default(),
         };
 
-        let provider_tool_config = ProviderToolCallConfig::from(&tool_config);
+        let provider_tool_config = tool_config.clone();
         let request = ModelInferenceRequest {
             inference_id: Uuid::now_v7(),
             messages: vec![RequestMessage {
@@ -1387,7 +1376,7 @@ mod tests {
         // Verify tools
         let tools = tools.unwrap();
         assert_eq!(tools.len(), 1);
-        assert_eq!(tools[0].function.name, WEATHER_TOOL.name());
+        assert_eq!(tools[0].function.name, WEATHER_TOOL_DEF.name);
 
         // Verify tool_choice is Specific
         let tool_choice = tool_choice.unwrap();
@@ -1830,7 +1819,7 @@ mod tests {
 
     #[test]
     fn test_mistral_apply_inference_params_called() {
-        let logs_contain = crate::utils::testing::capture_logs();
+        let logs_contain = crate::test_helpers::capture_logs();
         let inference_params = ChatCompletionInferenceParamsV2 {
             reasoning_effort: Some("high".to_string()),
             service_tier: None,
@@ -2044,7 +2033,8 @@ mod tests {
 
     #[test]
     fn test_mistral_assistant_message_with_thought_blocks() {
-        use crate::inference::types::{ContentBlock, Text, Thought};
+        use tensorzero_inference_types::ContentBlock;
+        use tensorzero_types::{Text, Thought};
 
         let message = RequestMessage {
             role: Role::Assistant,
