+++ crates/tensorzero-providers/src/xai.rs
@@ -12,39 +12,38 @@ use tensorzero_types_providers::xai::{
 use tokio::time::Instant;
 use url::Url;
 
-use crate::endpoints::inference::InferenceCredentials;
-use crate::error::{
-    DelayedError, DisplayOrDebugGateway, Error, ErrorDetails, warn_discarded_thought_block,
-};
-use crate::http::TensorzeroHttpClient;
-use crate::inference::InferenceProvider;
-use crate::inference::types::batch::{BatchRequestRow, PollBatchInferenceResponse};
-use crate::inference::types::chat_completion_inference_params::{
-    ChatCompletionInferenceParamsV2, warn_inference_parameter_not_supported,
-};
-use crate::inference::types::usage::raw_usage_entries_from_value;
-use crate::inference::types::{
-    ApiType, ContentBlock, ContentBlockOutput, Latency, ModelInferenceRequest,
-    ModelInferenceRequestJsonMode, PeekableProviderInferenceResponseStream,
-    ProviderInferenceResponse, ProviderInferenceResponseArgs, ProviderInferenceResponseChunk,
-    ProviderInferenceResponseStreamInner, RequestMessage, Role, Text, Thought, Unknown,
-    batch::StartBatchProviderInferenceResponse,
-};
-use crate::model::Credential;
-use crate::model::{ModelProviderRequestInfo, ProviderInferenceRequest};
-use crate::providers::helpers::{
+use crate::helpers::{
     inject_extra_request_data_and_send, inject_extra_request_data_and_send_eventsource,
 };
-use crate::providers::openai::{
+use crate::openai::{
     OpenAIMessagesConfig, OpenAIResponseChoice, ReasoningFieldName, StreamOptions, get_chat_url,
     handle_openai_error, openai_response_tool_call_to_tensorzero_tool_call, prepare_file_message,
     stream_openai,
 };
+use tensorzero_error::{
+    DelayedError, DisplayOrDebugGateway, Error, ErrorDetails, warn_discarded_thought_block,
+};
+use tensorzero_http::TensorzeroHttpClient;
+use tensorzero_inference_types::credentials::Credential;
+use tensorzero_inference_types::credentials::{ModelProviderRequestInfo, ProviderInferenceRequest};
+use tensorzero_inference_types::provider_trait::InferenceProvider;
+use tensorzero_inference_types::raw_usage_entries_from_value;
+use tensorzero_inference_types::utils::warn_inference_parameter_not_supported;
+use tensorzero_inference_types::{BatchRequestRow, PollBatchInferenceResponse};
+use tensorzero_inference_types::{
+    ContentBlock, ContentBlockOutput, Latency, ModelInferenceRequest,
+    ModelInferenceRequestJsonMode, PeekableProviderInferenceResponseStream,
+    ProviderInferenceResponse, ProviderInferenceResponseArgs, ProviderInferenceResponseChunk,
+    ProviderInferenceResponseStreamInner, RequestMessage, StartBatchProviderInferenceResponse,
+};
+use tensorzero_types::inference_params::ChatCompletionInferenceParamsV2;
+use tensorzero_types::inference_params::InferenceCredentials;
+use tensorzero_types::{ApiType, Role, Text, Thought, Unknown};
 use uuid::Uuid;
 
-use crate::inference::TensorZeroEventError;
-use crate::providers::chat_completions::prepare_chat_completion_tools;
-use crate::providers::chat_completions::{ChatCompletionTool, ChatCompletionToolChoice};
+use crate::chat_completions::prepare_chat_completion_tools;
+use crate::chat_completions::{ChatCompletionTool, ChatCompletionToolChoice};
+use tensorzero_inference_types::provider_trait::TensorZeroEventError;
 
 lazy_static! {
     static ref XAI_DEFAULT_BASE_URL: Url = {
@@ -897,17 +896,14 @@ mod tests {
 
     use super::*;
 
-    use crate::inference::types::{
-        FinishReason, FunctionType, ModelInferenceRequestJsonMode, RequestMessage, Role,
-    };
-    use crate::providers::chat_completions::{
+    use crate::chat_completions::{
         ChatCompletionSpecificToolChoice, ChatCompletionSpecificToolFunction,
         ChatCompletionToolChoice, ChatCompletionToolType,
     };
-    use crate::providers::openai::{
-        OpenAIFinishReason, OpenAIResponseChoice, OpenAIResponseMessage,
-    };
-    use crate::providers::test_helpers::{WEATHER_PROVIDER_TOOL_CONFIG, WEATHER_TOOL};
+    use crate::openai::{OpenAIFinishReason, OpenAIResponseChoice, OpenAIResponseMessage};
+    use crate::test_helpers::{WEATHER_PROVIDER_TOOL_CONFIG, WEATHER_TOOL_DEF};
+    use tensorzero_inference_types::{FinishReason, ModelInferenceRequestJsonMode, RequestMessage};
+    use tensorzero_types::{FunctionType, Role};
     use tensorzero_types_providers::xai::XAIUsage;
 
     #[tokio::test]
@@ -948,15 +944,15 @@ mod tests {
         assert_eq!(tools.len(), 1);
 
         let tool = &tools[0];
-        assert_eq!(tool.function.name, WEATHER_TOOL.name());
-        assert_eq!(tool.function.parameters, WEATHER_TOOL.parameters());
+        assert_eq!(tool.function.name, WEATHER_TOOL_DEF.name);
+        assert_eq!(tool.function.parameters, &WEATHER_TOOL_DEF.parameters);
         assert_eq!(
             xai_request.tool_choice,
             Some(ChatCompletionToolChoice::Specific(
                 ChatCompletionSpecificToolChoice {
                     r#type: ChatCompletionToolType::Function,
                     function: ChatCompletionSpecificToolFunction {
-                        name: WEATHER_TOOL.name(),
+                        name: &WEATHER_TOOL_DEF.name,
                     }
                 }
             ))
@@ -1002,15 +998,15 @@ mod tests {
         assert_eq!(tools.len(), 1);
 
         let tool = &tools[0];
-        assert_eq!(tool.function.name, WEATHER_TOOL.name());
-        assert_eq!(tool.function.parameters, WEATHER_TOOL.parameters());
+        assert_eq!(tool.function.name, WEATHER_TOOL_DEF.name);
+        assert_eq!(tool.function.parameters, &WEATHER_TOOL_DEF.parameters);
         assert_eq!(
             xai_request.tool_choice,
             Some(ChatCompletionToolChoice::Specific(
                 ChatCompletionSpecificToolChoice {
                     r#type: ChatCompletionToolType::Function,
                     function: ChatCompletionSpecificToolFunction {
-                        name: WEATHER_TOOL.name(),
+                        name: &WEATHER_TOOL_DEF.name,
                     }
                 }
             ))
@@ -1125,7 +1121,7 @@ mod tests {
 
     #[test]
     fn test_xai_apply_inference_params_called() {
-        let logs_contain = crate::utils::testing::capture_logs();
+        let logs_contain = crate::test_helpers::capture_logs();
         let inference_params = ChatCompletionInferenceParamsV2 {
             reasoning_effort: Some("high".to_string()),
             service_tier: None,
